+++ changelog.md
@@ -1,3 +1,6 @@
+## 2022-10-05
+- minor bugfixes regarding Worklist.save and Sample.fit
+
 ## 2022-10-04
 - fixed bug introduced by the use of pathlib.Path
 - cleaned up settings.ini